opt_requirements.txt
# Optional requirements:
sqlalchemy==1.4.0
python-dotenv
